html/browsers/browsing-the-web/navigating-across-documents/child_navigates_parent_location.html

<!doctype html>
<!--/
specs:
  - id: browsing-the-web.html#navigating-across-documents
  - id: browsers.html#nested-browsing-context

description: |
  1. Creates an iframe with src="child_navigates_parent_location-1.html"
  2. child_navigates_parent_location-1.html sends the message "initial" and
      creates an iframe with src="child_navigates_parent_location-2.html"
  3. child_navigates_parent_location-2.html sends the message "inner" and
      then sets its parent's location (the iframe with
      src="child_navigates_parent_location-1.html") to
      "child_navigates_parent_location-3.html"
  4. child_navigates_parent_location-3.html sends the message "destination"

/-->
<title>Child document navigating parent via location</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var test = async_test();
var reports = [];
onmessage = test.step_func(function(e) {
  reports.push(e.data);
  if (reports.length == 3) {
    assert_array_equals(reports, ["initial", "inner", "destination"]);
    test.done();
  }
});
</script>
<iframe src="child_navigates_parent_location-1.html"></iframe>
